+++ frontend/src/components/layout/navbar.tsx
@@ -1,10 +1,14 @@
-// import { MobileSidebar } from './mobile-sidebar';
+'use client';
+
 import { UserNav } from './user-nav';
 import Link from 'next/link';
 import { ThemeToggler } from '../theme-toggler';
 import { Sun } from 'lucide-react';
+import { cn } from '@/lib/utils';
+import { SidebarMobile } from './sidebar-mobile';
+import { Account } from '../../../../infrastructure/graphql/api/codegen/appsync';
 
-export default function Navbar() {
+export default function Navbar({ accounts }: { accounts: [Account] }) {
   return (
     <div className='supports-backdrop-blur:bg-background/60 fixed left-0 right-0 top-0 z-20 border-b bg-background/95 backdrop-blur'>
       <nav className='flex h-12 items-center justify-between px-4'>
@@ -25,9 +29,9 @@ export default function Navbar() {
             <span>Pecuniary</span>
           </Link>
         </div>
-        {/* <div className={cn('block lg:!hidden')}>
-          <MobileSidebar />`
-        </div> */}
+        <div className={cn('block lg:!hidden')}>
+          <SidebarMobile accounts={accounts} />
+        </div>
         <div className='flex items-center gap-1 px-3'>
           <Sun className='w-6 h-6' />
           <ThemeToggler />
