--- frontend/src/components/layout/sidebar-wrapper.tsx
@@ -1,10 +1,0 @@
-'use server';
-
-import Sidebar from '@/components/layout/sidebar';
-import { fetchAccounts } from '@/actions';
-
-export default async function SidebarWrapper() {
-  const accounts = await fetchAccounts();
-
-  return <Sidebar accounts={accounts} />;
-}
